QuantumMereology

Documentation for QuantumMereology.

QuantumMereology.GOE

QuantumMereology.optimize_spectral

QuantumMereology.GOE — Method

GOE(N::Integer)

Traceless Gaussian Orthogonal Ensemble (GOE) random matrix of size 2^N x 2^N.

QuantumMereology.optimize_spectral — Method

optimize_spectral(H::AbstractMatrix, strings::Vector{<:SparseMatrixCSC}, iterations::Int; verbose=false)

Find a unitary U such that U*H*U' is supported on the given set of Pauli strings.

Arguments

H::AbstractMatrix: the input Hamiltonian matrix.

strings::Vector{<:SparseMatrixCSC}: the set of Pauli strings defining the target support.

iterations::Int: number of optimization iterations.

verbose=false: if true, print progress information.

Returns

U: a unitary matrix such that U*H*U' is supported on strings.

Details

Minimizes the cost function ||E - E0||², where E are the eigenvalues of U*H*U' and E0 are the eigenvalues of H.

References

https://www.pnas.org/doi/10.1073/pnas.2308006120